ewc-continual-learning/requirements.txt
numpy==1.24.3  # Numerical operations and array handling
torch==2.1.0  # Continual learning model and EWC
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
pytest==7.4.0  # Testing framework
pytest-cov==4.1.0  # Test coverage reporting

